Moved the custom context to its own file and added a new core package that is used to request a timeout context and it handles the switch between simulated context and real context based on if the passed LocalClock is castable to a *SimClock
