+++ core/sync/sync.go
@@ -2,6 +2,7 @@ package sync
 
 import (
 	"context"
+	"example.com/scion-time/core/contextcore"
 	"example.com/scion-time/simulation/simutils"
 	"time"
 
@@ -70,8 +71,7 @@ func RegisterClocks(refClocks, netClocks []client.ReferenceClock) *SyncableClock
 
 func (c *SyncableClocks) measureOffsetToRefClocks(log *zap.Logger, lclk timebase.LocalClock, timeout time.Duration) time.Duration {
 	log.Debug("Measuring offset to reference clocks", zap.String("clock holder id", c.Id))
-	//ctx, cancel := context.WithDeadline(context.Background(), lclk.Now().Add(timeout))
-	ctx, cancel := simutils.WithDeadline(lclk, timeout) // TODO this is still breaking for normal use
+	ctx, cancel := contextcore.WithTimeout(lclk, context.Background(), timeout)
 	defer cancel()
 	c.refClkClient.MeasureClockOffsets(ctx, log, c.refClks, c.refClkOffsets)
 	return timemath.Median(c.refClkOffsets)
@@ -125,8 +125,7 @@ func RunLocalClockSync(log *zap.Logger, lclk timebase.LocalClock, syncClks *Sync
 
 func (c *SyncableClocks) measureOffsetToNetClocks(log *zap.Logger, lclk timebase.LocalClock, timeout time.Duration) time.Duration {
 	log.Debug("\033[47mMeasuring offset to net clocks\033[0m", zap.String("clock holder id", c.Id), zap.Int("num net clks", len(c.netClks)))
-	//ctx, cancel := context.WithDeadline(context.Background(), lclk.Now().Add(timeout))
-	ctx, cancel := simutils.WithDeadline(lclk, timeout) // TODO this is still breaking for normal use
+	ctx, cancel := contextcore.WithTimeout(lclk, context.Background(), timeout)
 	defer cancel()
 	c.netClkClient.MeasureClockOffsets(ctx, log, c.netClks, c.netClkOffsets)
 	log.Debug("\033[47mFinished measuring offset to net clocks\033[0m", zap.String("clock holder id", c.Id))
